[#18] Reverted absolute imports

+++ src/contexts/log/LogProvider.tsx
@@ -1,33 +1,33 @@
 import { addDoc, collection, deleteDoc, doc, DocumentData, getDocs, getFirestore } from 'firebase/firestore';
 import { PropsWithChildren, useCallback, useEffect, useState } from 'react';
 
-import { auth, firebaseApp } from '@/config';
+import { auth, firebaseApp } from '../../config';
 
 import { LogContext } from './LogContext';
-import type { Log, LogMessageId, RawLog } from './';
+import { Log, RawLog } from '../../hooks';
 
-const db = getFirestore(firebaseApp);
-const logsCollection = collection(db, 'logs');
+const db = firebaseApp ? getFirestore(firebaseApp) : null;
+const logsCollection = db ? collection(db, 'logs') : null;
 
 export const LogProvider = ({ children }: PropsWithChildren) => {
 	const [logs, updateLogs] = useState<DocumentData[]>([]);
 
 	const refreshData = useCallback(async () => {
-		const logDBData = await getDocs(logsCollection);
-		const updatedLogs = logDBData.docs.map(doc => {
+		const logDBData = logsCollection ? await getDocs(logsCollection) : null;
+		const updatedLogs = logDBData ? logDBData.docs.map(doc => {
 			return {
 				...doc.data(),
 				id: doc.id,
 			} as Log;
-		});
+		}) : [];
 
 		orderBy('newFirst', updatedLogs);
 		updateLogs(updatedLogs);
 	}, []);
 
 	const deleteLog = useCallback(
-		async (id: LogMessageId) => {
-			if (id) {
+		async (id: string) => {
+			if (db && id) {
 				const logDoc = doc(db, `logs/${id}`);
 
 				await deleteDoc(logDoc);
@@ -39,22 +39,25 @@ export const LogProvider = ({ children }: PropsWithChildren) => {
 
 	const addLog = useCallback(
 		async (data: RawLog): Promise<boolean> => {
-			try {
-				const timestamp = new Date().toISOString();
-				const newLog = {
-					...data,
-					timestamp,
-					creator: auth.currentUser?.uid || '',
-				};
-
-				await addDoc(logsCollection, newLog);
-				await refreshData();
-
-				return true;
-			} catch (error) {
-				console.error('Error trying to fetch log DB', error);
-				return false;
+			if (logsCollection) {
+				try {
+					const timestamp = new Date().toISOString();
+					const newLog = {
+						...data,
+						timestamp,
+						creator: auth?.currentUser?.uid || '',
+					};
+	
+					await addDoc(logsCollection, newLog);
+					await refreshData();
+	
+					return true;
+				} catch (error) {
+					console.error('Error trying to fetch log DB', error);
+					return false;
+				}
 			}
+			return false;
 		},
 		[refreshData],
 	);
